requirements.txt
discord.py==2.0.0
python-dotenv==1.0.0
paramiko==2.11.0
